Refactor DWARF add2line
