+++ pkg/symbol/elfutils/elfutils.go
@@ -33,6 +33,7 @@ var dwarfSuffix = func(s *elf.Section) string {
 	}
 }
 
+// HasDWARF reports whether the specified executable or library file contains DWARF debug information.
 func HasDWARF(path string) (bool, error) {
 	f, err := elf.Open(path)
 	if err != nil {
@@ -71,6 +72,7 @@ func readableDWARFSections(f *elf.File) (map[string]struct{}, error) {
 	return exists, nil
 }
 
+// IsSymbolizableGoObjFile checks whether the specified executable or library file is generated by Go toolchain.
 func IsSymbolizableGoObjFile(path string) (bool, error) {
 	// Checks ".note.go.buildid" section and symtab better to keep those sections in object file.
 	f, err := elf.Open(path)
